Scripts/requirements.txt
pynput>=1.7.6
pillow>=10.0.0
pyautogui>=0.9.54
opencv-python>=4.8.0
numpy>=1.24.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
mss>=9.0.1
winrt>=1.0.0
qrcode[pil]>=7.4.2
edge-tts>=6.1.0
appopener>=1.0.0
pygame>=2.5.0
websockets>=12.0
aiortc>=1.7.0
av>=12.0.0
pywin32>=306
